1. useGetCurrentPosition custom hook 추가 2. api map type -> global type 변경 3. Google Map region 분기처리 로직 변경 4. gas -> oil 명칭 변경
