+++ src/components/Map/index.ts
@@ -3,5 +3,5 @@ export {default as GoogleMap} from './GoogleMap';
 export {default as SearchBox} from './SearchBox';
 export {default as Swtich} from './Swtich';
 export {default as ListButton} from './ListButton';
-export {default as GasStationMarker} from './GasStationMarker';
+export {default as OilStationMarker} from './OilStationMarker';
 export {default as MapZoomPanel} from './MapZoomPanel';
