+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export {default as useKakaoAuthActions} from './useKakaoAuthActions';
 export {default as useKakaoUser} from './useKakaoUser';
 export {default as useGetColumnGutter} from './useGetColumnGutter';
+export {default as useGetCurrentPosition} from './useGetCurrentPosition';
